+++ test/backends/typst/decorations.jl
@@ -15,6 +15,16 @@
   table(
     align: (right, right, right,),
     columns: (auto, auto, auto,),
+    stroke: none,
+    // == Horizontal Lines =================================================================
+    table.hline(y: 0, stroke: 1.5pt,),
+    table.hline(y: 1, stroke: 0.8pt,),
+    table.hline(y: 4, stroke: 1.5pt,),
+    // == Vertical Lines ===================================================================
+    table.vline(x: 0, end: 4, stroke: 1.5pt),
+    table.vline(x: 1, end: 4, stroke: 0.8pt),
+    table.vline(x: 2, end: 4, stroke: 0.8pt),
+    table.vline(x: 3, end: 4, stroke: 1.5pt),
     // == Table Header =====================================================================
     table.header(
       // -- Column Labels: Row 1 -----------------------------------------------------------
@@ -54,6 +64,16 @@
   table(
     align: (right, right, right,),
     columns: (auto, auto, auto,),
+    stroke: none,
+    // == Horizontal Lines =================================================================
+    table.hline(y: 0, stroke: 1.5pt,),
+    table.hline(y: 1, stroke: 0.8pt,),
+    table.hline(y: 4, stroke: 1.5pt,),
+    // == Vertical Lines ===================================================================
+    table.vline(x: 0, end: 4, stroke: 1.5pt),
+    table.vline(x: 1, end: 4, stroke: 0.8pt),
+    table.vline(x: 2, end: 4, stroke: 0.8pt),
+    table.vline(x: 3, end: 4, stroke: 1.5pt),
     // == Table Header =====================================================================
     table.header(
       // -- Column Labels: Row 1 -----------------------------------------------------------
@@ -98,6 +118,16 @@
   table(
     align: (right, right, right,),
     columns: (auto, auto, auto,),
+    stroke: none,
+    // == Horizontal Lines =================================================================
+    table.hline(y: 0, stroke: 1.5pt,),
+    table.hline(y: 2, stroke: 0.8pt,),
+    table.hline(y: 5, stroke: 1.5pt,),
+    // == Vertical Lines ===================================================================
+    table.vline(x: 0, end: 5, stroke: 1.5pt),
+    table.vline(x: 1, end: 5, stroke: 0.8pt),
+    table.vline(x: 2, end: 5, stroke: 0.8pt),
+    table.vline(x: 3, end: 5, stroke: 1.5pt),
     // == Table Header =====================================================================
     table.header(
       // -- Column Labels: Row 1 -----------------------------------------------------------
@@ -138,10 +168,6 @@
         )
 
         @test result == expected
-
-
-        expected = """
-"""
     end
 
     @testset "Table Property Validation" verbose = true begin