+++ test/backends/typst/offset_arrays.jl
@@ -20,6 +20,17 @@
   table(
     align: (right, right, right, right,),
     columns: (auto, auto, auto, auto,),
+    stroke: none,
+    // == Horizontal Lines =================================================================
+    table.hline(y: 0, stroke: 1.5pt,),
+    table.hline(y: 1, stroke: 0.8pt,),
+    table.hline(y: 4, stroke: 1.5pt,),
+    // == Vertical Lines ===================================================================
+    table.vline(x: 0, end: 4, stroke: 1.5pt),
+    table.vline(x: 1, end: 4, stroke: 0.8pt),
+    table.vline(x: 2, end: 4, stroke: 0.8pt),
+    table.vline(x: 3, end: 4, stroke: 0.8pt),
+    table.vline(x: 4, end: 4, stroke: 1.5pt),
     // == Table Header =====================================================================
     table.header(
       // -- Column Labels: Row 1 -----------------------------------------------------------